rename legacy references to queue to more appropriate worker terminology
