+++ plugin-server/src/main/jobs/local/mock-graphile-worker.ts
@@ -67,7 +67,7 @@ export class MockGraphileWorker {
         }
     }
 
-    async startConsumer(jobHandlers: TaskList): Promise<void> {
+    async start(jobHandlers: TaskList): Promise<void> {
         this.jobHandlers = jobHandlers
         if (!this.started) {
             this.started = true
@@ -100,7 +100,7 @@ export class MockGraphileWorker {
         return false
     }
 
-    async stopConsumer(): Promise<void> {
+    async stop(): Promise<void> {
         this.started = false
         await this.syncState()
         fs.unlinkSync(this.filename)
@@ -113,12 +113,12 @@ export class MockGraphileWorker {
         }
     }
 
-    async pauseConsumer(): Promise<void> {
+    async pause(): Promise<void> {
         this.paused = true
         await this.syncState()
     }
 
-    isConsumerPaused(): boolean {
+    isPaused(): boolean {
         return this.paused
     }
 }
