+++ plugin-server/tests/main/capabilities.test.ts
@@ -2,7 +2,7 @@ import Piscina from '@posthog/piscina'
 
 import { KafkaQueue } from '../../src/main/ingestion-queues/kafka-queue'
 import { startQueues } from '../../src/main/ingestion-queues/queue'
-import { startJobsConsumer } from '../../src/main/jobs/job-queue-consumer'
+import { startJobsConsumer } from '../../src/main/jobs/worker-setup'
 import { Hub, LogLevel } from '../../src/types'
 import { createHub } from '../../src/utils/db/hub'
 
@@ -47,37 +47,37 @@ describe('capabilities', () => {
 
     describe('startJobsConsumer()', () => {
         it('sets up bufferJob handler if ingestion is on', async () => {
-            hub.graphileWorker.startConsumer = jest.fn()
+            hub.graphileWorker.start = jest.fn()
             hub.capabilities.ingestion = true
             hub.capabilities.processPluginJobs = false
 
             await startJobsConsumer(hub, piscina)
 
-            expect(hub.graphileWorker.startConsumer).toHaveBeenCalledWith({
+            expect(hub.graphileWorker.start).toHaveBeenCalledWith({
                 bufferJob: expect.anything(),
             })
         })
 
         it('sets up pluginJob handler if processPluginJobs is on', async () => {
-            hub.graphileWorker.startConsumer = jest.fn()
+            hub.graphileWorker.start = jest.fn()
             hub.capabilities.ingestion = false
             hub.capabilities.processPluginJobs = true
 
             await startJobsConsumer(hub, piscina)
 
-            expect(hub.graphileWorker.startConsumer).toHaveBeenCalledWith({
+            expect(hub.graphileWorker.start).toHaveBeenCalledWith({
                 pluginJob: expect.anything(),
             })
         })
 
         it('sets up bufferJob and pluginJob handlers if ingestion and processPluginJobs are on', async () => {
-            hub.graphileWorker.startConsumer = jest.fn()
+            hub.graphileWorker.start = jest.fn()
             hub.capabilities.ingestion = true
             hub.capabilities.processPluginJobs = true
 
             await startJobsConsumer(hub, piscina)
 
-            expect(hub.graphileWorker.startConsumer).toHaveBeenCalledWith({
+            expect(hub.graphileWorker.start).toHaveBeenCalledWith({
                 bufferJob: expect.anything(),
                 pluginJob: expect.anything(),
             })
